cli-impl ― ColumbaFunctionalTest refactor.
