+++ .github/workflows/publish-plugin.yaml
@@ -26,7 +26,7 @@ jobs:
           distribution: zulu
           java-version: 17.0.6
 
-      - name: Performs 'publish' task.
+      - name: Performs 'publish-plugin' task.
         uses: gradle/gradle-build-action@v2
         with:
-          arguments: publish
+          arguments: publish-plugin
